+++ test/integration/rmq_2_sysmon/monitor_queue.py
@@ -17,7 +17,11 @@
 # Standard
 import sys
 import os
-import unittest
+
+if sys.version_info < (2, 7):
+    import unittest2 as unittest
+else:
+    import unittest
 
 # Third-party
 import mock
@@ -86,13 +90,7 @@ class UnitTest(unittest.TestCase):
         rmq_2_sysmon.monitor_queue(self.cfg, self.log)
         self.log.log_close()
 
-        if self.connect_true in open(self.cfg.log_file).read():
-            status = True
-
-        else:
-            status = False
-
-        self.assertTrue(status)
+        self.assertTrue(self.connect_true in open(self.cfg.log_file).read())
 
     def tearDown(self):
 
@@ -105,7 +103,8 @@ class UnitTest(unittest.TestCase):
         """
 
         os.remove(self.cfg.log_file)
-        rmq_cleanup.rmq_cleanup(self.cfg, self.cfg.queue_name, True)
+        rmq_cleanup.rmq_cleanup(self.cfg, self.cfg.queue_list[0]["queue"],
+                                True)
 
 
 if __name__ == "__main__":